Separate out builtin function collection into its own class
This makes for a better separation between the declarations and the logic that defines builtins.

# Conflicts:
#	libyul/backends/evm/EVMDialect.cpp
