+++ libyul/CMakeLists.txt
@@ -58,6 +58,8 @@ add_library(yul
 	backends/evm/ControlFlowGraphBuilder.h
 	backends/evm/EthAssemblyAdapter.cpp
 	backends/evm/EthAssemblyAdapter.h
+	backends/evm/EVMBuiltins.cpp
+	backends/evm/EVMBuiltins.h
 	backends/evm/EVMCodeTransform.cpp
 	backends/evm/EVMCodeTransform.h
 	backends/evm/EVMDialect.cpp
